Refactor: Update BasicViewModel and BasicNavViewModel
The `replay` parameter for `MutableSharedFlow` in both `BasicViewModel` and `BasicNavViewModel` has been changed from `1` to `0`. This means that new subscribers will not receive the last emitted value.

The `extraBufferCapacity` for `intent` in `BasicViewModel` has been changed from `3` to `1`. In `BasicNavViewModel`, it has been changed from `3` to `2`. The `extraBufferCapacity` for `uiMessageIntent` in both ViewModels has been changed from `3` to `1`.

The nullable types for `basicIntent` and `basicUiMessageIntent` in `MutableSharedFlow` have been changed to non-nullable types.

The `clearIntentFlow` and `clearUiMessageIntentFlow` methods have been removed from both ViewModels.

A new abstract class `BasicSimpleViewModel` has been added. This class extends `ViewModel` and has an abstract method `setEvent`.