Refactor: Extract `ReceiptIntent` handling into a separate function
The logic for handling `ReceiptIntent`s in `ReceiptCompose` has been extracted into a new private function `handleReceiptIntent`. This improves readability and organization of the `ReceiptCompose` composable.

Additionally, `ExperimentalMaterial3Api` annotation was added and the navigation logic for `GoToSplitReceiptScreen` and `GoToEditReceiptsScreen` was updated to set `inclusive = true` and `saveState` (for split screen) when popping up to `CreateReceiptScreenNav`. The `clearIntentFlow` call was also removed as it's handled within the `setEvent` call.