Adjust `OrderReport` font size and add `translatedName` to preview
The font size for the "no_order_report" text in `OrderReport` has been changed from 24.sp to 20.sp and the font weight set to normal.

In the `SplitReceiptScreenPreview`, the `orderReportText` is now set to null and a `translatedName` has been added to one of the `OrderData` items for testing purposes.