Limit maximum amount of orders
Added a limit to the maximum number of orders that can be added to a receipt.
The limit is set to `DataConstantsReceipt.MAXIMUM_AMOUNT_OF_DISHES`.
A new state `isOrderCountAtLimit` is introduced to indicate whether the limit has been reached.
The `monitorAmountOfOrders` function is used to update this state based on the current number of orders.