feat: Implement `ChooseFolderNameBottomSheet`
This commit introduces a new bottom sheet dialog, `ChooseFolderNameBottomSheet`, for selecting a folder name.

The dialog displays a list of unarchived folders, allowing users to choose one.
It utilizes `ModalBottomSheet` for the presentation and `FlowGridLayout` to display folder names.
A preview for the `ChooseFolderNameView` is also included.

Additionally, a minor refactor in `ChooseLanguageDialog.kt` renames `DialogCap` to `DialogCapView` for consistency.